Edit: edited locale file from locales.config to locale.yml Edit: Edited dependCheck and added detection of Di Anti-cheat. Package-Change: Edited package name from commands to Commands. Delete: deleted localesFailException.java Delete: deleted loadInternalLocales.java
