--- src/main/java/org/wasteutils/WasteUtils/Comands/CommandAnnouncement.java
@@ -1,4 +1,0 @@
-package org.wasteutils.WasteUtils.Comands;
-
-public class CommandAnnouncement {
-}
