--- src/main/java/org/wasteutils/WasteUtils/Comands/CommandMail.java
@@ -1,4 +1,0 @@
-package org.wasteutils.WasteUtils.Comands;
-
-public class CommandMail {
-}
