+++ src/main/java/org/wasteutils/WasteUtils/Commands/CommandAnnouncement.java
@@ -1,0 +1,4 @@
+package org.wasteutils.WasteUtils.Commands;
+
+public class CommandAnnouncement {
+}
