+++ src/main/java/org/wasteutils/WasteUtils/Commands/CommandMail.java
@@ -1,0 +1,4 @@
+package org.wasteutils.WasteUtils.Commands;
+
+public class CommandMail {
+}
