+++ src/main/java/org/wasteutils/WasteUtils/Commands/CommandManager.java
@@ -1,4 +1,4 @@
-package org.wasteutils.WasteUtils.Comands;
+package org.wasteutils.WasteUtils.Commands;
 
 /*
 Commands manager
